Grazie Signore

+++ app/src/main/java/it/polito/kgame/PesoInfo.kt
@@ -3,6 +3,6 @@ package it.polito.kgame
 
 data class PesoInfo(
         var data: Long? = null,
-        var peso : Long? = null,
+        var peso : Float? = null,
 
 )
